Add files via upload
I sometimes forget the basic syntax of reading and writing from file. There might be some people looking for simple and basic syntax for the same purpose. So i thought i should upload this code on git so that it may help others. Better versions with better logic may be available on some other branch. It was written back in February 2015 uploaded in 2020.